Adding New UserListViewModel for SwiftUI
