+++ RandomUsers/Presentation/UserScene/UserList/ViewModel/UserListViewModel.swift
@@ -98,53 +98,54 @@ final class DefaultUserListViewModel: UserListViewModel, ObservableObject {
             page += 1
         }
         
-        self.userListUseCase
-            .execute(requestValue: .init(query: UserQuery.init(gender: self.gender.getQueryString(), page: "\(page)")))
-            .sink { completion in
-                switch completion {
-                case .finished:
-                    self.isLoading = false
-                    print("Finished userList viewModel call")
-                case .failure(let error):
-                    print("Failed userList viewModel call : ", error)
-                }
-            } receiveValue: { result in
-                switch result {
-                case .success(let data):
-                    data.toDomain().users.forEach {
-                        if !self.usersUUID.contains($0.loginInfo.uuid) {
-                            self.users.append($0)
-                        }
-                    }
-                    
-                    completion(nil)
-                case .failure(let parsingError):
-                    print("error : ", parsingError)
-                }
-            }
-            .store(in: &subscriptions)
+//        self.userListUseCase
+//            .execute(requestValue: .init(query: UserQuery.init(gender: self.gender.getQueryString(), page: "\(page)")))
+//            .sink { completion in
+//                switch completion {
+//                case .finished:
+//                    self.isLoading = false
+//                    print("Finished userList viewModel call")
+//                case .failure(let error):
+//                    print("Failed userList viewModel call : ", error)
+//                }
+//            } receiveValue: { result in
+//                switch result {
+//                case .success(let data):
+//                    data.toDomain().users.forEach {
+//                        if !self.usersUUID.contains($0.loginInfo.uuid) {
+//                            self.users.append($0)
+//                        }
+//                    }
+//                    
+//                    completion(nil)
+//                case .failure(let parsingError):
+//                    print("error : ", parsingError)
+//                }
+//            }
+//            .store(in: &subscriptions)
 
         
         
-//        self.userListUseCase.execute(
-//            requestValue: .init(query: UserQuery.init(gender: self.gender.getQueryString(), page: "\(page)"))
-//        ) { [weak self] response in
-//            guard let self = self else { return }
-//            
-//            self.isLoading = false
-//
-//            switch response {
-//            case .success(let page):
-//                page.users.forEach {
-//                    if !self.usersUUID.contains($0.loginInfo.uuid) {
-//                        self.users.append($0)
-//                    }
-//                }
-//                completion(nil)
-//            case .failure(let error):
-//                completion(error)
-//            }
-//        }
+        self.userListUseCase.execute(
+            requestValue: .init(query: UserQuery.init(gender: self.gender.getQueryString(), page: "\(page)"))
+        ) { [weak self] response in
+            guard let self = self else { return }
+            
+            self.isLoading = false
+
+            switch response {
+            case .success(let page):
+                page.users.forEach {
+                    if !self.usersUUID.contains($0.loginInfo.uuid) {
+                        self.users.append($0)
+                        self.usersUUID.append($0.loginInfo.uuid)
+                    }
+                }
+                completion(nil)
+            case .failure(let error):
+                completion(error)
+            }
+        }
     }
     
     func resetPage() {
